+++ include/em_duration.h
@@ -4,87 +4,97 @@
 #include "em_defs.h"
 #include "em_threading.h"
 
+template<typename T, typename I> class EmDuration_;
+
+// The default duration that supports a maximum of 49 days and 17 hours time.
+using EmDuration = EmDuration_<uint32_t, ts_uint32>;
+
+// A short duration object in case you need to handle shorter durations up to 65 seconds.
+using EmDurationShort = EmDuration_<uint16_t, ts_uint16>;
+
+// The long duration object in case you need to handle longer durations.
+using EmDurationLong = EmDuration_<uint64_t, ts_uint64>;
+
 // EmDuration class for handling duration-related operations
-// NOTE:
-// due to 32 bit implementation a maximum duration of 49 days and 17 hours is supported!
-class EmDuration {
+template<typename T, typename I>
+class EmDuration_ {
 protected:
-    ts_uint32 m_durationMillis;  
+    I m_durationMillis;  
 
 public:
-    EmDuration(uint16_t hours,
-               uint16_t minutes, 
-               uint16_t seconds, 
-               uint16_t milliseconds=0) :
+    EmDuration_(uint16_t hours,
+                uint16_t minutes, 
+                uint16_t seconds, 
+                uint16_t milliseconds=0) :
         m_durationMillis( 
-            (static_cast<uint32_t>(hours) * 3600 * 1000) + 
-            (static_cast<uint32_t>(minutes) * 60 * 1000) + 
-            (static_cast<uint32_t>(seconds) * 1000) +       
-             static_cast<uint32_t>(milliseconds)) {}
+            (static_cast<T>(hours) * 3600 * 1000) + 
+            (static_cast<T>(minutes) * 60 * 1000) + 
+            (static_cast<T>(seconds) * 1000) +       
+             static_cast<T>(milliseconds)) {}
     
     // Using 'explicit' prevents unintentional conversions from integer types.
-    explicit EmDuration(uint32_t milliseconds) :
+    explicit EmDuration_(T milliseconds) :
         m_durationMillis(milliseconds) {}
 
-    EmDuration(const EmDuration& other) :
-        m_durationMillis(static_cast<uint32_t>(other.m_durationMillis)) {}
+    EmDuration_(const EmDuration_& other) :
+        m_durationMillis(static_cast<T>(other.m_durationMillis)) {}
 
     // Operators
-    bool operator ==(const EmDuration& other) const {
+    bool operator ==(const EmDuration_& other) const {
         return m_durationMillis == other.m_durationMillis;
     }
     
-    bool operator !=(const EmDuration& other) const {
+    bool operator !=(const EmDuration_& other) const {
         return m_durationMillis != other.m_durationMillis;
     }
     
-    bool operator <(const EmDuration& other) const {
+    bool operator <(const EmDuration_& other) const {
         return m_durationMillis < other.m_durationMillis;
     }
     
-    bool operator >(const EmDuration& other) const {
+    bool operator >(const EmDuration_& other) const {
         return m_durationMillis > other.m_durationMillis;
     }
     
-    bool operator <=(const EmDuration& other) const {
+    bool operator <=(const EmDuration_& other) const {
         return m_durationMillis <= other.m_durationMillis;
     }
     
-    bool operator >=(const EmDuration& other) const {
+    bool operator >=(const EmDuration_& other) const {
         return m_durationMillis >= other.m_durationMillis;
     }   
 
-    EmDuration operator +(uint32_t milliseconds) const {
-        return EmDuration(milliseconds + m_durationMillis);}
+    EmDuration_ operator +(T milliseconds) const {
+        return EmDuration_(milliseconds + m_durationMillis);}
 
-    EmDuration operator +(const EmDuration& other) const {
-        return EmDuration(m_durationMillis + other.m_durationMillis);
+    EmDuration_ operator +(const EmDuration_& other) const {
+        return EmDuration_(m_durationMillis + other.m_durationMillis);
     }
     
-    EmDuration operator -(uint32_t milliseconds) const {
-        return EmDuration(m_durationMillis - milliseconds);
+    EmDuration_ operator -(T milliseconds) const {
+        return EmDuration_(m_durationMillis - milliseconds);
     }
 
-    EmDuration operator -(const EmDuration& other) const {
-        return EmDuration(m_durationMillis - other.m_durationMillis);
+    EmDuration_ operator -(const EmDuration_& other) const {
+        return EmDuration_(m_durationMillis - other.m_durationMillis);
     }
 
-    EmDuration& operator +=(uint32_t milliseconds) {
+    EmDuration_& operator +=(T milliseconds) {
         m_durationMillis += milliseconds;
         return *this;
     }
     
-    EmDuration& operator +=(const EmDuration& other) {
+    EmDuration_& operator +=(const EmDuration_& other) {
         m_durationMillis += other.m_durationMillis;
         return *this;
     }
     
-    EmDuration& operator -=(uint32_t milliseconds) {
+    EmDuration_& operator -=(T milliseconds) {
         m_durationMillis -= milliseconds;
         return *this;
     }
     
-    EmDuration& operator -=(const EmDuration& other) {
+    EmDuration_& operator -=(const EmDuration_& other) {
         m_durationMillis -= other.m_durationMillis;
         return *this;
     }
@@ -105,7 +115,7 @@ public:
     }
 
     // Get the duration in milliseconds
-    uint32_t milliseconds() const {
+    T milliseconds() const {
         return m_durationMillis;
     }
 
@@ -114,7 +124,7 @@ public:
             uint16_t& minutes,
             uint16_t& seconds,
             uint16_t& milliseconds) const {
-        uint32_t totalMillis = m_durationMillis;
+        T totalMillis = m_durationMillis;
         days = static_cast<uint16_t>(totalMillis / (24 * 3600 * 1000));
         totalMillis %= (24 * 3600 * 1000);
         hours = static_cast<uint16_t>(totalMillis / (3600 * 1000));
@@ -128,7 +138,7 @@ public:
             uint16_t& minutes,
             uint16_t& seconds,
             uint16_t& milliseconds) const {
-        uint32_t totalMillis = m_durationMillis;
+        T totalMillis = m_durationMillis;
         hours = static_cast<uint16_t>(totalMillis / (3600 * 1000));
         totalMillis %= (3600 * 1000);
         minutes = static_cast<uint16_t>(totalMillis / (60 * 1000));